Move helper RRID functions from utils to rrid
This is needed to prevent hidden circular dependency
